+++ lib/src/model/ripple.dart
@@ -1,0 +1,9 @@
+import 'package:flutter/material.dart';
+
+class RippleModel {
+  final bool enable;
+  final Color highlightColor;
+  final Color splashColor;
+
+  RippleModel({this.enable, this.highlightColor, this.splashColor});
+}